+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 <!-- insert new changelog below this comment -->
 
+- feat(scripts): add SPECIFY_INIT_DIR to target a member project from the repo root (#2892)
+
 ## [0.11.2] - 2026-06-18
 
 ### Changed
@@ -1823,4 +1825,3 @@
 ### Changed
 
 - Update release.yml
-